+++ jitsi-meet/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Script error fixes with introduction turnserver
 * Adjust config.js and videobridge settings
 * Remove no-loopback-peers from turnserver config, bad config option
+* Allow setting default resolution as parameter
 
 ---